requirements.txt
aiogram==3.3.0
SQLAlchemy==2.0.23
aiosqlite==0.19.0
python-dotenv==1.0.0
APScheduler==3.10.4
